Complete order flow and minus quantity in stock after complete paymet is set in status

+++ milkstore/src/main/java/com/cookswp/milkstore/pojo/entities/Order.java
@@ -22,10 +22,10 @@ public class Order implements  Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private long id;
 
     @Column(name = "user_id", nullable = false)
-    private Long userId;
+    private int userId;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "order_status", nullable = false)
